+++ docs/CHANGELOG.md
@@ -13,7 +13,7 @@ Remove the `fastAssemble` task which is subsumed by `assembleForJavac`.
 **Closed issues:**
 
 
-Version 3.30.1 (March 1, 2023)
+Version 3.31.0 (March 1, 2023)
 ------------------------------
 
 **User-visible changes:**